Enhance deleteBranch method with existence check

+++ src/main/java/com.pranav.BMS.Service/BranchService.java
@@ -54,11 +54,16 @@ public class BranchService {
 	}
 
 	public ResponseEntity<ResponseStructure<Branch>> deleteBranch(String id) {
-		branchDAO.deleteBranch(id);
-		ResponseStructure<Branch> rs = new ResponseStructure<Branch>();
-		rs.setData(null);
-		rs.setMessage("Branch having id " + id + " deleted successfully");
-		rs.setStatusCode(HttpStatus.ACCEPTED.value());
-		return new ResponseEntity<ResponseStructure<Branch>>(rs, HttpStatus.ACCEPTED);
+		Branch data = branchDAO.getBranch(id);
+		if (data != null) {
+			branchDAO.deleteBranch(id);
+			ResponseStructure<Branch> rs = new ResponseStructure<Branch>();
+			rs.setData(null);
+			rs.setMessage("Branch having id " + id + " deleted successfully");
+			rs.setStatusCode(HttpStatus.ACCEPTED.value());
+			return new ResponseEntity<ResponseStructure<Branch>>(rs, HttpStatus.ACCEPTED);
+		} else {
+			throw new IdDoesNotPresentException("Branch id " + id + " not found");
+		}
 	}
 }
